requirements.txt
numpy==1.19.5
streamlit==1.0.0
matplotlib==3.4.3
seaborn==0.11.2
pandas==1.3.3
Pillow==8.3.2
scikit_learn==1.0
